pyproject.toml
[project]
name = "breathe-bot"
version = "0.1.0"
description = "Breathing assistant for discord."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "pre-commit>=4.2.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = true

[tool.ty.environment]
root = ["./src"]

[tool.ty.src]
exclude = ["docs"]
